[mesh] move Texturing::remapVisibilities -> Mesh::remapVisibilities
